redis2/requirements.txt
streamlit==1.28.0
haystack-ai==2.16.1
elasticsearch==8.10.0
pandas==2.1.3
requests==2.31.0
redis==5.0.1
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
aiofiles==23.2.1
python-dotenv==1.0.0
asyncpg==0.29.0
redis[hiredis]==5.0.1
pyyaml==6.0.1
